refactor(engine): split offscreen.rs into offscreen/{mask,blur,blit} submodules (move-only)
`offscreen.rs` (1504 LOC) mixed three GPU effect concerns — shader-mask, blur,
and surface-blit — in one `impl OffscreenRenderer`. Split move-only into an
`offscreen/` directory (mirrors the painter/ #286 and replay/ #287 splits):

  offscreen/mod.rs   (585) struct + BlitPipeline/BlurPipelines/MaskedRenderResult/
                           BlurParams/FullscreenVertex types + constructors + all
                           create_* resource helpers + warmup + accessors + tests
  offscreen/mask.rs  (278) get_or_create_pipeline + render_masked
  offscreen/blur.rs  (434) get_or_create_blur_pipelines + render_blur
  offscreen/blit.rs  (256) get_or_create_blit_pipeline + blit_to_surface

The submodules are descendant modules of `offscreen`, so their inherent
`impl OffscreenRenderer` blocks retain access to the struct's private fields.
One visibility bump: `get_or_create_pipeline` → `pub(super)` (it is called by
`warmup` in the parent). Inline `super::shader_compiler` rebased to
`super::super::shader_compiler` for the one-level-deeper modules; `DeviceExt`
re-imported in `blit.rs`. No other behaviour/signature change.

(Note: offscreen.rs was already C1-compliant at ~1427 non-test LOC; this is a
concern-separation/reviewability split, not a C1 fix — it completes the
god-module-split theme alongside painter/ and replay/.)

Pure move: zero behaviour change, proven by the full enable-wgpu-tests GPU
readback (444 passed, 0 failed) staying bit-identical.
